Fixed Combinations() method so can find all k combinations
Fixed bug in Combinations() method that was preventing it from working
when k!=2. Also, added a helper function to be used w/ this method.
Code was essentially copied w/ light modification to work w/ List<Gate>
instead of a int[] from an answer on stack overflow listed in
documentation for the method.

Also, removed original Combinations() method completely in favor of
this one.