Fixed bug causing gates with outdegree > 1 to not be partitioned correctly
Looking in standard viewing direction of a graph (inputs —> outputs),
gates w/ outdegree > 1 were not being partitioned properly. This is b/c
the path splitter wasn’t adding each half of the path to a subgraph
when it encountered a qs node (it was only adding path terminating in
qs node).