Refactored PartitionCircuit() class
Completely refactored PartitionCircuit() class in an attempt to
implement the “strategy” design pattern. Moved most of the algorithmic
code out of PartitionCircuit and into a separate class that subclasses
PartitionCircuit and implements the PartitionCircuitAlgorithm interface.

The rationale behind the subclassing is that the algorithm requires the
Subgraph helper object and this is the easiest way to provide access.
In the future, may want to change Subgraph so it’s not a nested class
(but it’s also pretty ugly and usage doesn’t really make sense out of
the context of PartitionCircuit). If Subgraph is no longer a nested
class, PartitionCircuitPathPartition() no longer needs to subclass
PartitionCircuit() itself. It’s currently just a sort of hacky
workaround.